+++ lib/flycal_cli/pipeline.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "flycal_cli/pipeline/params"
+require "flycal_cli/pipeline/retriever"
+require "flycal_cli/pipeline/aggregator"
+require "flycal_cli/pipeline/renderer"
+require "flycal_cli/pipeline/text_renderer"
+require "flycal_cli/pipeline/search_pipeline"
+
+module FlycalCli
+  module Pipeline
+  end
+end